clean up. restore support for blocking after recent chagnes
